Move live client contract out of protocol
